Add Adobe stand-alone Flash Player
Icon from https://helpx.adobe.com/flash-player.html

Note that the Flatpak has (read-only) access to all of the filesystem
as well as xdg-run/gvfs so that it can access the data files in the same
directory as the SWF file we're trying to open. Most Flash games and demos
are not self-contained, and will require access to data files in the
directory or deeper into the hierarchy.